refactor(modals): move components/modals → sections/modals, adopt formik
Move all modals from web/src/components/modals/ to web/src/sections/modals/
and update all import sites.

FormField upgrades:
- CreateProjectModal: replace manual useState with Formik + InputTypeInField
  + Yup validation (required, trimmed)
- AddInstructionModal: replace manual useState with Formik + InputTextAreaField
- EditPropertyModal: replace legacy TextFormField from @/components/Field
  with InputTypeInField + InputVertical layout

Fix oxfmt pre-commit hook to strip the web/ path prefix before passing
filenames so paths resolve correctly after cd web.
