+++ .pre-commit-config.yaml
@@ -147,7 +147,7 @@ repos:
 
       - id: oxfmt
         name: oxfmt
-        entry: bash -c 'cd web && bunx oxfmt "$@"' _
+        entry: bash -c 'cd web && bunx oxfmt "${@#web/}"' _
         language: system
         types_or: [html, css, javascript, ts, tsx]
