+++ web/src/app/admin/bots/SlackBotUpdateForm.tsx
@@ -9,7 +9,7 @@ import { SlackTokensForm } from "./SlackTokensForm";
 
 import { EditableStringFieldDisplay } from "@/components/EditableStringFieldDisplay";
 import { deleteSlackBot } from "./new/lib";
-import GenericConfirmModal from "@/components/modals/GenericConfirmModal";
+import GenericConfirmModal from "@/sections/modals/GenericConfirmModal";
 import { Button } from "@opal/components";
 import { cn } from "@opal/utils";
 import { SvgChevronDownSmall, SvgTrash } from "@opal/icons";
