--- web/src/components/modals/AddInstructionModal.tsx
@@ -1,57 +1,0 @@
-"use client";
-
-import { useEffect, useState } from "react";
-import { Button } from "@opal/components";
-import { useProjectsContext } from "@/providers/ProjectsContext";
-import InputTextArea from "@/refresh-components/inputs/InputTextArea";
-import { useModal } from "@/refresh-components/contexts/ModalContext";
-import { SvgAddLines } from "@opal/icons";
-import Modal from "@/refresh-components/Modal";
-
-export default function AddInstructionModal() {
-  const modal = useModal();
-  const { currentProjectDetails, upsertInstructions } = useProjectsContext();
-  const [instructionText, setInstructionText] = useState("");
-
-  useEffect(() => {
-    if (!modal.isOpen) return;
-    const preset = currentProjectDetails?.project?.instructions ?? "";
-    setInstructionText(preset);
-  }, [modal.isOpen, currentProjectDetails?.project?.instructions]);
-
-  async function handleSubmit() {
-    const value = instructionText.trim();
-    try {
-      await upsertInstructions(value);
-    } catch (e) {
-      console.error("Failed to save instructions", e);
-    }
-    modal.toggle(false);
-  }
-
-  return (
-    <Modal open={modal.isOpen} onOpenChange={modal.toggle}>
-      <Modal.Content width="sm">
-        <Modal.Header
-          icon={SvgAddLines}
-          title="Set Project Instructions"
-          description="Specify the behaviors or tone for the chat sessions in this project."
-          onClose={() => modal.toggle(false)}
-        />
-        <Modal.Body>
-          <InputTextArea
-            value={instructionText}
-            onChange={(event) => setInstructionText(event.target.value)}
-            placeholder="My goal with is to... be sure to... in your responses."
-          />
-        </Modal.Body>
-        <Modal.Footer>
-          <Button prominence="secondary" onClick={() => modal.toggle(false)}>
-            Cancel
-          </Button>
-          <Button onClick={handleSubmit}>Save Instructions</Button>
-        </Modal.Footer>
-      </Modal.Content>
-    </Modal>
-  );
-}
