+++ web/src/refresh-pages/admin/ImageGenerationPage/forms/ImageGenFormWrapper.tsx
@@ -2,7 +2,7 @@
 
 import { useState, useMemo, useEffect } from "react";
 import { Form, Formik, FormikProps } from "formik";
-import ProviderModal from "@/components/modals/ProviderModal";
+import ProviderModal from "@/sections/modals/ProviderModal";
 import ModelIcon from "@/app/admin/configuration/language-models/ModelIcon";
 import ConnectionProviderIcon from "@/refresh-components/ConnectionProviderIcon";
 import {
