+++ web/src/sections/modals/ConfirmEntityModal.tsx
